requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via httpx
asttokens==3.0.1
    # via stack-data
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cmdstanpy==1.3.0
    # via prophet
colorama==0.4.6
    # via
    #   ipython
    #   tqdm
contourpy==1.3.3
    # via matplotlib
cryptography==46.0.3
    # via pyjwt
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via ipython
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
dotenv==0.9.9
    # via -r requirements.in
executing==2.2.1
    # via stack-data
fonttools==4.61.0
    # via matplotlib
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
holidays==0.86
    # via prophet
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-resources==6.5.2
    # via prophet
ipython==9.8.0
    # via -r requirements.in
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
joblib==1.5.2
    # via scikit-learn
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.7
    # via
    #   -r requirements.in
    #   prophet
    #   seaborn
matplotlib-inline==0.2.1
    # via
    #   -r requirements.in
    #   ipython
multidict==6.7.0
    # via yarl
numpy==2.3.5
    # via
    #   -r requirements.in
    #   cmdstanpy
    #   contourpy
    #   matplotlib
    #   pandas
    #   patsy
    #   prophet
    #   ruptures
    #   scikit-learn
    #   scipy
    #   seaborn
    #   stanio
    #   statsmodels
    #   xgboost
packaging==25.0
    # via
    #   deprecation
    #   matplotlib
    #   statsmodels
pandas==2.3.3
    # via
    #   -r requirements.in
    #   cmdstanpy
    #   prophet
    #   seaborn
    #   statsmodels
parso==0.8.5
    # via jedi
patsy==1.0.2
    # via statsmodels
pillow==12.0.0
    # via matplotlib
postgrest==2.25.0
    # via supabase
prompt-toolkit==3.0.52
    # via ipython
propcache==0.4.1
    # via yarl
prophet==1.2.1
    # via -r requirements.in
psycopg2==2.9.11
    # via -r requirements.in
pure-eval==0.2.3
    # via stack-data
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   postgrest
    #   realtime
    #   storage3
    #   supabase-auth
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
pyjwt[crypto]==2.10.1
    # via supabase-auth
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   holidays
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via dotenv
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via -r requirements.in
realtime==2.25.0
    # via supabase
requests==2.32.5
    # via -r requirements.in
ruptures==1.1.10
    # via -r requirements.in
scikit-learn==1.8.0
    # via -r requirements.in
scipy==1.16.3
    # via
    #   ruptures
    #   scikit-learn
    #   statsmodels
    #   xgboost
seaborn==0.13.2
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
stack-data==0.6.3
    # via ipython
stanio==0.5.1
    # via cmdstanpy
statsmodels==0.14.6
    # via -r requirements.in
storage3==2.25.0
    # via supabase
strenum==0.4.15
    # via supabase-functions
supabase==2.25.0
    # via -r requirements.in
supabase-auth==2.25.0
    # via supabase
supabase-functions==2.25.0
    # via supabase
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via
    #   cmdstanpy
    #   prophet
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   anyio
    #   pydantic
    #   pydantic-core
    #   realtime
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.6.1
    # via requests
wcwidth==0.2.14
    # via prompt-toolkit
websockets==15.0.1
    # via realtime
xgboost==3.1.2
    # via -r requirements.in
yarl==1.22.0
    # via
    #   postgrest
    #   storage3
    #   supabase-functions
